Move SSL classes into their own package
